requirements.txt
python == 3.7.6
nltk >= 3.4.3
numpy >= 1.14
pytorch == 1.4.0